Fix build warning by moving types to their own files
