--- src/MessagePack.UnityClient/Assets/Scripts/MessagePack/Resolvers/UnsafeBinaryResolver.cs
@@ -1,102 +1,0 @@
-﻿// Copyright (c) All contributors. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System;
-using MessagePack.Formatters;
-using MessagePack.Internal;
-
-#pragma warning disable SA1403 // File may only contain a single namespace
-
-namespace MessagePack.Resolvers
-{
-    public sealed class NativeGuidResolver : IFormatterResolver
-    {
-        /// <summary>
-        /// The singleton instance that can be used.
-        /// </summary>
-        public static readonly NativeGuidResolver Instance = new NativeGuidResolver();
-
-        private NativeGuidResolver()
-        {
-        }
-
-        public IMessagePackFormatter<T> GetFormatter<T>()
-        {
-            return FormatterCache<T>.Formatter;
-        }
-
-        private static class FormatterCache<T>
-        {
-            public static readonly IMessagePackFormatter<T> Formatter;
-
-            static FormatterCache()
-            {
-                Formatter = (IMessagePackFormatter<T>)NativeGuidResolverGetFormatterHelper.GetFormatter(typeof(T));
-            }
-        }
-    }
-
-    public sealed class NativeDecimalResolver : IFormatterResolver
-    {
-        /// <summary>
-        /// The singleton instance that can be used.
-        /// </summary>
-        public static readonly NativeDecimalResolver Instance = new NativeDecimalResolver();
-
-        private NativeDecimalResolver()
-        {
-        }
-
-        public IMessagePackFormatter<T> GetFormatter<T>()
-        {
-            return FormatterCache<T>.Formatter;
-        }
-
-        private static class FormatterCache<T>
-        {
-            public static readonly IMessagePackFormatter<T> Formatter;
-
-            static FormatterCache()
-            {
-                Formatter = (IMessagePackFormatter<T>)NativeDecimalResolverGetFormatterHelper.GetFormatter(typeof(T));
-            }
-        }
-    }
-}
-
-namespace MessagePack.Internal
-{
-    internal static class NativeGuidResolverGetFormatterHelper
-    {
-        internal static object GetFormatter(Type t)
-        {
-            if (t == typeof(Guid))
-            {
-                return NativeGuidFormatter.Instance;
-            }
-            else if (t == typeof(Guid?))
-            {
-                return new StaticNullableFormatter<Guid>(NativeGuidFormatter.Instance);
-            }
-
-            return null;
-        }
-    }
-
-    internal static class NativeDecimalResolverGetFormatterHelper
-    {
-        internal static object GetFormatter(Type t)
-        {
-            if (t == typeof(Decimal))
-            {
-                return NativeDecimalFormatter.Instance;
-            }
-            else if (t == typeof(Decimal?))
-            {
-                return new StaticNullableFormatter<Decimal>(NativeDecimalFormatter.Instance);
-            }
-
-            return null;
-        }
-    }
-}
